Enable CmdLetBinding and better handle re-execution
CmdLetBinding allows the script to be called with the parameter
`-ErrorAction "Stop"`, which exits as soon as an error occurs. This
allows wrapper scripts to detect and handle errors.

Support for re-execution allows the script to be called multiple times
on the same system, and still result in a working configuration.